% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumGain.R
\name{cumGain}
\alias{cumGain}
\title{Calculates the cumulative gain between sequential values.}
\usage{
cumGain(x)
}
\arguments{
\item{x}{A numeric vector of values.}
}
\value{
Returns a single value sum of all gains.
}
\description{
Calculate the sum of every gain in a series, i.e., for the series x = 4, 3, 2, 5, 6 the cumulative gain is 3 + 1 = 4.
}
\details{
cumGain returns the cumulative gain between consecutive values in a series.
}
\examples{
x <- c(400, 500, 500, 450, 550)  # Some example elevation values
cumGain(x)                       # Calc. cum gain (AKA tot. ascent)
}
\author{
Bjorn J. Brooks
}
