% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumLoss.R
\name{cumLoss}
\alias{cumLoss}
\title{Calculates the cumulative loss between sequential values.}
\usage{
cumLoss(x)
}
\arguments{
\item{x}{A numeric vector of values.}
}
\value{
Returns a single value sum of all losses.
}
\description{
Calculate the sum of every loss in a series, i.e., for the
  series x = 4, 3, 2, 5, 6 the cumulative loss is 1 + 1 = 2.
}
\details{
cumLoss returns the cumulative loss between consecutive values in a
  series.
}
\examples{
x <- c(400, 500, 500, 450, 550)  # Some example elevation values
cumLoss(x)                       # Calc. cum loss (AKA tot. descent)
}
\author{
Bjorn J. Brooks
}
