% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustDist.R
\name{hclustDist}
\alias{hclustDist}
\title{Wrapper for hierarchical clustering of spatial coordinates.}
\usage{
hclustDist(xy, method = "ward.D", dS = NULL)
}
\arguments{
\item{xy}{A numeric array of spatial coordinates.}

\item{method}{A character string specifying the hierarchical clustering
method to be passed to stats::hclust.}

\item{dS}{An optional numeric value specifying the denominator to be used in
down-scaling each value of input xy.}
}
\value{
Returns an object of class *hclust*, see stats::hclust.
}
\description{
Convenient wrapper for hierarchical clustering spatial
  coordinates.
}
\details{
hclustDist returns results from hierarchical clustering using the
  stats::hclust() function.
}
\examples{
xy <- data.frame(x=c(0, 0, 2, 3),               # Some example coordinates
                 y=c(0, 1, 4, 5))
rownames(xy) <- LETTERS[1:4]                    # Set names of coord points
fit <- hclustDist(xy)

# Plot
par(mfrow=c(1,2))
plot(xy, pch='')
text(xy, labels=rownames(xy))

plot(fit)
}
\author{
Bjorn J. Brooks
}
